requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
gql==0.1.0
graphql-core==2.1
idna==2.7
Jinja2==2.10
MarkupSafe==1.1.0
promise==2.2.1
PyYAML==3.13
requests==2.20.1
Rx==1.6.1
six==1.11.0
urllib3==1.24.1
